+++ config/config.go
@@ -3,9 +3,9 @@ package config
 import (
 	"encoding/json"
 	"errors"
+	"github.com/Sirupsen/logrus"
 	log "github.com/Sirupsen/logrus"
 	"os"
-	"github.com/Sirupsen/logrus"
 )
 
 type Config struct {
@@ -33,7 +33,7 @@ type Logger struct {
 
 type DB struct {
 	Implementation string `json:"implementation"`
-	Connection string `json:"connection"`
+	Connection     string `json:"connection"`
 }
 
 type Statsd struct {
