+++ handlers/registerNode.go
@@ -7,8 +7,8 @@ import (
 	"strconv"
 
 	log "github.com/Sirupsen/logrus"
-	"github.com/qa-dev/jsonwire-grid/pool"
 	"github.com/qa-dev/jsonwire-grid/jsonwire"
+	"github.com/qa-dev/jsonwire-grid/pool"
 )
 
 type RegisterNode struct {
