+++ storage/mysql_test.go
@@ -6,9 +6,9 @@ import (
 	"fmt"
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/jmoiron/sqlx"
-	"github.com/rubenv/sql-migrate"
 	"github.com/qa-dev/jsonwire-grid/pool"
 	"github.com/qa-dev/jsonwire-grid/storage/mysql"
+	"github.com/rubenv/sql-migrate"
 	"os"
 	"strings"
 	"testing"
