+++ utils/mocks/node/main.go
@@ -1,22 +1,20 @@
 package main
 
 import (
-	log "github.com/Sirupsen/logrus"
-	"net/http"
-	"os/exec"
-	"github.com/qa-dev/jsonwire-grid/jsonwire"
-	"encoding/json"
-	"regexp"
-	"strings"
-	"flag"
-	"time"
-	"net"
-	"strconv"
 	"bytes"
-	"io/ioutil"
-	"fmt"
+	"encoding/json"
 	"errors"
+	"flag"
+	"fmt"
+	log "github.com/Sirupsen/logrus"
+	"github.com/qa-dev/jsonwire-grid/jsonwire"
+	"io/ioutil"
 	"math/rand"
+	"net"
+	"net/http"
+	"strconv"
+	"strings"
+	"time"
 )
 
 var currentSessionID, host, hubUrl string
@@ -30,8 +28,8 @@ func main() {
 	}
 
 	hubUrlFlag := flag.String("hub", "http://127.0.0.1:4444", "address of hub, default http://127.0.0.1:4444")
-	rand.Seed( time.Now().UTC().UnixNano())
-	portFlag := flag.Int("port", rand.Intn(1000) + 5000, "port default, rand")
+	rand.Seed(time.Now().UTC().UnixNano())
+	portFlag := flag.Int("port", rand.Intn(1000)+5000, "port default, rand")
 	maxDurationFlag := flag.Int("maxDuration", 0, "request duration [0 <=duration], default 0")
 	flag.Parse()
 	hubUrl = *hubUrlFlag
@@ -40,11 +38,12 @@ func main() {
 	log.Infof("hub url: %v", hubUrl)
 	log.Infof("port: %v", port)
 	log.Infof("maxDuration: %v", maxDuration)
+
 	register()
 
 	go func() {
 		for {
-			<- time.Tick(time.Second)
+			<-time.Tick(time.Second)
 			err := sendApiProxy()
 			if err != nil {
 				log.Errorf("Error send [api/proxy], ", err)
@@ -56,13 +55,13 @@ func main() {
 	http.HandleFunc("/wd/hub/session/", useSession)
 	http.HandleFunc("/wd/hub/sessions", getSessions)
 
-	err = http.ListenAndServe(":" + strconv.Itoa(port), nil)
+	err = http.ListenAndServe(":"+strconv.Itoa(port), nil)
 	if err != nil {
 		log.Errorf("Listen serve error, %s", err)
 	}
 }
 
-// getIpv4 метод получает ip адрес, на котором стартует данный сервис.
+// getIpv4 get ip of this server
 func getIpv4() (string, error) {
 	addrs, err := net.InterfaceAddrs()
 	if err != nil {
@@ -79,74 +78,11 @@ func getIpv4() (string, error) {
 	return "", nil
 }
 
-func createSession(rw http.ResponseWriter, r *http.Request) {
-	if maxDuration > 0 {
-		rand.Seed( time.Now().UTC().UnixNano())
-		time.Sleep(time.Millisecond * time.Duration(rand.Intn(maxDuration)))
-	}
-	rw.Header().Set("Accept", "application/json")
-	rw.Header().Set("Accept-charset", "utf-8")
-
-	if r.Method != http.MethodPost {
-		http.Error(rw, "Method Not Allowed", http.StatusMethodNotAllowed)
-		return
-	}
-	if currentSessionID != "" {
-		errorMassage := "Session already exists"
-		log.Error(errorMassage)
-		rw.WriteHeader(http.StatusInternalServerError)
-		responseMessage := &jsonwire.Message{}
-		responseMessage.Status = int(jsonwire.RESPONSE_STATUS_UNKNOWN_ERR)
-		responseMessage.Value = errorMassage
-		json.NewEncoder(rw).Encode(responseMessage)
-		return
-	}
-
-	out, err := exec.Command("uuidgen").Output()
-	if err != nil {
-		log.Fatalf("Can't generate uuid, %s", err)
-	}
-	currentSessionID = string(out[:len(out) - 1]) // cut end of line char
-	json.NewEncoder(rw).Encode(&jsonwire.Message{SessionId: currentSessionID})
-}
-
-func useSession(rw http.ResponseWriter, r *http.Request) {
-	if maxDuration > 0 {
-		rand.Seed( time.Now().UTC().UnixNano())
-		time.Sleep(time.Millisecond * time.Duration(rand.Intn(maxDuration)))
-	}
-	rw.Header().Set("Accept", "application/json")
-	rw.Header().Set("Accept-charset", "utf-8")
-
-	re := regexp.MustCompile(".*/session/([^/]+)(?:/([^/]+))?")
-	parsedUrl := re.FindStringSubmatch(r.URL.Path)
-	if len(parsedUrl) != 3 {
-		errorMessage := "url [" + r.URL.Path + "] parsing error"
-		log.Infof(errorMessage)
-		http.Error(rw, errorMessage, http.StatusBadRequest)
-		return
-	}
-	sessionId := re.FindStringSubmatch(r.URL.Path)[1]
-	responseMessage := &jsonwire.Message{SessionId: sessionId}
-	if sessionId != currentSessionID {
-		errorMassage := fmt.Sprintf("sessionID '%s' not found", sessionId)
-		log.Error(errorMassage)
-		rw.WriteHeader(http.StatusNotFound)
-		responseMessage.Status = int(jsonwire.RESPONSE_STATUS_UNKNOWN_ERR)
-		responseMessage.Value = errorMassage
-		json.NewEncoder(rw).Encode(responseMessage)
-		return
-	}
-	if parsedUrl[2] == "" && r.Method == http.MethodDelete {
-		currentSessionID = ""
-	}
-	json.NewEncoder(rw).Encode(responseMessage)
-}
-
+// register register on hub
 func register() {
 	log.Info("Try register")
 	register := jsonwire.Register{
-		Configuration: &jsonwire.Configuration{Host:host, Port: port},
+		Configuration:    &jsonwire.Configuration{Host: host, Port: port},
 		CapabilitiesList: []jsonwire.Capabilities{{"browserName": "firefox"}},
 	}
 	b, err := json.Marshal(register)
@@ -155,7 +91,7 @@ func register() {
 		return
 	}
 
-	req, err := http.NewRequest(http.MethodPost, hubUrl + "/grid/register", bytes.NewBuffer(b))
+	req, err := http.NewRequest(http.MethodPost, hubUrl+"/grid/register", bytes.NewBuffer(b))
 	if err != nil {
 		log.Errorf("Can't register, create request error, %s", err)
 		return
@@ -192,9 +128,10 @@ func register() {
 
 }
 
+// sendApiProxy check "is server know me" and register if server return false
 func sendApiProxy() error {
 	b := strings.NewReader("{}")
-	req, err := http.NewRequest(http.MethodPost, hubUrl + "/grid/api/proxy?id=http://" + host + ":" + strconv.Itoa(port) , b)
+	req, err := http.NewRequest(http.MethodPost, hubUrl+"/grid/api/proxy?id=http://"+host+":"+strconv.Itoa(port), b)
 	if err != nil {
 		err = errors.New(fmt.Sprintf("create request error, %s", err))
 		return err
@@ -236,30 +173,3 @@ func sendApiProxy() error {
 	}
 	return nil
 }
-
-func getSessions(rw http.ResponseWriter, r *http.Request) {
-	sessions := &jsonwire.Sessions{}
-	if currentSessionID != "" {
-		sessions.Value = []struct {
-			Id           string `json:"id"`
-			Capabilities json.RawMessage `json:"capabilities"`
-		}{
-			{Id: currentSessionID, Capabilities: nil},
-		}
-	}
-
-
-	err := json.NewEncoder(rw).Encode(sessions)
-	if err != nil {
-		err = errors.New("Get sessions error, " + err.Error())
-		log.Error(err)
-		json.NewEncoder(rw).Encode(&jsonwire.Message{Value: err.Error(), Status: int(jsonwire.RESPONSE_STATUS_UNKNOWN_ERR)})
-	}
-
-
-
-}
-
-
-
-
